+++ blob/lib/highlevel.node.ts
@@ -15,10 +15,10 @@ import { IBlobAccessConditions } from "./models";
 import { Batch } from "./utils/Batch";
 import { BufferScheduler } from "./utils/BufferScheduler";
 import {
-  BLOB_DEFAULT_DOWNLOAD_BLOCK_BYTES,
   BLOCK_BLOB_MAX_BLOCKS,
   BLOCK_BLOB_MAX_STAGE_BLOCK_BYTES,
-  BLOCK_BLOB_MAX_UPLOAD_BLOB_BYTES
+  BLOCK_BLOB_MAX_UPLOAD_BLOB_BYTES,
+  DEFAULT_BLOB_DOWNLOAD_BLOCK_BYTES
 } from "./utils/constants";
 import { generateBlockID } from "./utils/utils.common";
 import { streamToBuffer } from "./utils/utils.node";
@@ -101,13 +101,13 @@ async function uploadResetableStreamToBlockBlob(
     );
   }
   if (options.blockSize === 0) {
-    if (size > BLOCK_BLOB_MAX_STAGE_BLOCK_BYTES * BLOCK_BLOB_MAX_BLOCKS) {
+    if (size > BLOCK_BLOB_MAX_BLOCKS * BLOCK_BLOB_MAX_STAGE_BLOCK_BYTES) {
       throw new RangeError(`${size} is too larger to upload to a block blob.`);
     }
     if (size > BLOCK_BLOB_MAX_UPLOAD_BLOB_BYTES) {
       options.blockSize = Math.ceil(size / BLOCK_BLOB_MAX_BLOCKS);
-      if (options.blockSize < BLOB_DEFAULT_DOWNLOAD_BLOCK_BYTES) {
-        options.blockSize = BLOB_DEFAULT_DOWNLOAD_BLOCK_BYTES;
+      if (options.blockSize < DEFAULT_BLOB_DOWNLOAD_BLOCK_BYTES) {
+        options.blockSize = DEFAULT_BLOB_DOWNLOAD_BLOCK_BYTES;
       }
     }
   }
@@ -197,7 +197,7 @@ export async function downloadBlobToBuffer(
     throw new RangeError("blockSize option must be >= 0");
   }
   if (options.blockSize === 0) {
-    options.blockSize = BLOB_DEFAULT_DOWNLOAD_BLOCK_BYTES;
+    options.blockSize = DEFAULT_BLOB_DOWNLOAD_BLOCK_BYTES;
   }
 
   if (offset < 0) {
@@ -240,7 +240,8 @@ export async function downloadBlobToBuffer(
         off,
         chunkEnd - off + 1,
         {
-          blobAccessConditions: options.blobAccessConditions
+          blobAccessConditions: options.blobAccessConditions,
+          maxRetryRequests: options.maxRetryRequestsPerBlock
         }
       );
       const stream = response.readableStreamBody!;